User priviledges
Assign priviledge level required to be able to create new posts in and
reply to threads in certain subforums. Accordingly blocks users from
making such content. This commit is incompatible with database files
created in earlier commits (no migration script exists).

Note the example in config.default.toml for an example of how to assign
priviledges to a subforum.